lotto-31 Tranfered get_list_str_numbers() method from LottoParser to Game class.
